# openapi.yaml
openapi: 3.1.0
info:
  title: Synthetic Data Generation Service
  version: 1.0.0
  description: |
    REST API для сервиса генерации конфиденциальных синтетических табличных данных
    с применением дифференциальной приватности (DP-CTGAN).

    ## Основной сценарий использования

    1. Загрузить датасет — `POST /datasets`
    2. Загрузить конфигурацию — `POST /configs`
    3. Запустить пайплайн — `POST /runs`
    4. Опрашивать статус — `GET /runs/{run_id}` (или дождаться webhook)
    5. Скачать синтетику — `GET /runs/{run_id}/synthetic`
    6. Изучить отчёт — `GET /runs/{run_id}/report`

    ## Асинхронное выполнение

    Обучение DP-CTGAN занимает несколько минут. Все запуски пайплайна
    (`POST /runs`) возвращают `run_id` немедленно со статусом `queued`.
    Итоговый статус — `completed` | `failed` | `cancelled`.

    ## Авторизация

    Все запросы требуют заголовок `Authorization: Bearer <token>`.
  contact:
    name: Сергей Рогачев
    email: s.rogachev@example.com
  license:
    name: MIT
servers:
- url: http://localhost:8000/api/v1
  description: Локальная разработка
- url: https://synth.internal/api/v1
  description: Внутренний production-стенд
...
security:
- BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          examples:
          - NOT_FOUND
        message:
          type: string
          examples:
          - Запуск с указанным ID не найден
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          examples:
          - 42
        page:
          type: integer
          examples:
          - 1
        per_page:
          type: integer
          examples:
          - 20
        pages:
          type: integer
          examples:
          - 3
...
    ModelSummary:
      type: object
      properties:
        model_id:
          type: string
          format: uuid
          examples:
          - 7c9e6679-7425-40de-944b-e07fc1f90ae7
        name:
          type: string
          examples:
          - adult_epsilon3_v2
        run_id:
          type:
          - string
          - 'null'
          format: uuid
        dataset_name:
          type: string
          examples:
          - adult_census
        created_at:
          type: string
          format: date-time
        file_size_bytes:
          type: integer
          examples:
          - 4718592
        epsilon:
          type: number
          examples:
          - 3.0
        epochs_completed:
          type: integer
          examples:
          - 287
        spent_epsilon:
          type:
          - number
          - 'null'
          examples:
          - 2.49
    ModelDetail:
      allOf:
      - $ref: '#/components/schemas/ModelSummary'
      - type: object
        properties:
          dp_config:
            $ref: '#/components/schemas/GeneratorReport/properties/dp_config'
          dp_spent:
            $ref: '#/components/schemas/GeneratorReport/properties/dp_spent'
          sample_size:
            type: integer
            description: Размер обучающей выборки
            examples:
            - 39073
    SampleRequest:
      type: object
      required:
      - n_rows
      properties:
        n_rows:
          type: integer
          minimum: 1
          maximum: 1000000
          examples:
          - 5000
        output_format:
          type: string
          enum:
          - csv
          - json
          - parquet
          default: csv
...
  parameters:
...
    ModelId:
      name: model_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Уникальный идентификатор модели
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Невалидный запрос (ошибка валидации входных данных)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                message: n_rows должен быть > 0
    Unauthorized:
      description: Отсутствует или невалиден токен авторизации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHORIZED
                message: Bearer token обязателен
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: NOT_FOUND
                message: Запуск не найден
...
paths:
...
  /models:
    get:
      tags:
      - models
      summary: Список сохранённых моделей
      operationId: listModels
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - name: dataset_name
        in: query
        schema:
          type: string
        description: Фильтр по датасету
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /models/{model_id}:
    get:
      tags:
      - models
      summary: Метаданные модели
      description: Возвращает конфигурацию и DP-отчёт сохранённой модели без самих весов.
      operationId: getModel
      parameters:
      - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - models
      summary: Удалить модель
      description: Удаляет .pkl-файл и запись в registry. Синтетика, сгенерированная ранее, сохраняется.
      operationId: deleteModel
      parameters:
      - $ref: '#/components/parameters/ModelId'
      responses:
        '204':
          description: Модель удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /models/{model_id}/samples:
    post:
      tags:
      - models
      summary: Сгенерировать синтетику из модели
      description: |
        Генерирует синтетические строки из сохранённой модели без пайплайна.
        Обучение не запускается. Оценка приватности и полезности не производится.
        Используйте для быстрого получения синтетики из уже проверенной модели.
      operationId: sampleFromModel
      parameters:
      - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SampleRequest'
            examples:
              default:
                value:
                  n_rows: 10000
                  output_format: csv
      responses:
        '200':
          description: Файл синтетических данных
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...